requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.5.0
idna==3.10
loguru==0.7.2
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
pybit==5.8.0
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
yarl==1.16.0
